Goodness-of-fit test for a single FLE covariate

Implements the univariate cumulative martingale residual test of Boschi & Wit (2025), Section 3.3. The test statistic is \(T_x = \sup_u |\hat W[u]|\) where \(\hat W[u]\) is the normalised cumulative score process for the requested covariate; under correct specification \(\hat W\) converges to a standard Brownian bridge, so the p-value follows the Kolmogorov-Smirnov distribution \(2 \sum_{k\ge 1} (-1)^{k-1} e^{-2 k^2 t^2}\).

Usage

gof_univariate(
  event_log,
  model,
  covariate,
  scope = "all",
  mode = "one",
  half_life = NULL,
  seed = NULL
)

Arguments

event_log

Dyadic event log.

model

Named character vector of <stat> = "linear" mapping.

covariate

Name of the covariate in model to test.

scope, mode, half_life, seed

See compare_models().

Value

A list with statistic (\(T_x\)), p_value (KS), W (numeric vector of length n, the normalised process), and u (the time grid in [0, 1]).

References

Boschi M, Wit EC (2025). Goodness of fit in relational event models. Statistics and Computing 36(4).

Examples

# \donttest{
data(classroom_events)
gof_univariate(classroom_events,
  model = c(reciprocity_count  = "linear",
            transitivity_count = "linear"),
  covariate = "reciprocity_count", seed = 1)
#> $statistic
#> [1] 0.7495182
#> 
#> $p_value
#> [1] 0.6279781
#> 
#> $W
#>   [1] -1.400075e-15  1.459096e-02  1.459096e-02  2.918191e-02  2.918191e-02
#>   [6]  4.377287e-02  4.377287e-02  5.836383e-02  5.836383e-02  5.836383e-02
#>  [11]  5.836383e-02  5.836383e-02  5.836383e-02  5.836383e-02  3.098026e-02
#>  [16]  9.400354e-03  9.400354e-03 -1.217955e-02 -1.217955e-02 -1.217955e-02
#>  [21] -1.217955e-02 -1.217955e-02 -3.956312e-02 -2.775679e-02 -2.074736e-02
#>  [26] -4.813094e-02 -4.813094e-02 -4.813094e-02 -4.813094e-02 -4.813094e-02
#>  [31] -4.813094e-02 -4.813094e-02 -4.813094e-02 -4.813094e-02 -7.551451e-02
#>  [36] -7.551451e-02 -7.551451e-02 -7.551451e-02 -7.551451e-02 -6.824359e-02
#>  [41] -6.824359e-02 -6.824359e-02 -6.824359e-02 -6.824359e-02 -6.824359e-02
#>  [46] -5.643726e-02 -5.643726e-02 -4.184631e-02 -6.922988e-02 -5.742355e-02
#>  [51] -5.742355e-02 -4.561722e-02 -4.561722e-02 -3.381089e-02 -3.381089e-02
#>  [56] -1.921993e-02 -1.921993e-02 -1.194902e-02 -1.194902e-02 -1.194902e-02
#>  [61] -2.735068e-02 -2.735068e-02 -1.275972e-02 -1.275972e-02 -1.275972e-02
#>  [66] -1.041673e-02 -1.041673e-02 -1.041673e-02  1.389600e-03  2.211696e-02
#>  [71]  3.807935e-02  1.649944e-02  1.097782e-03 -6.706147e-03 -6.706147e-03
#>  [76] -6.706147e-03 -2.828605e-02 -2.828605e-02 -4.674860e-02 -4.674860e-02
#>  [81] -1.380927e-01 -1.380927e-01 -1.380927e-01 -1.342263e-01 -1.342263e-01
#>  [86] -1.342263e-01 -1.248751e-01 -1.535438e-01 -1.720063e-01 -1.602000e-01
#>  [91] -1.425900e-01 -1.332602e-01 -1.548401e-01 -1.548401e-01 -1.548401e-01
#>  [96] -1.548401e-01 -1.548401e-01 -1.548401e-01 -1.548401e-01 -1.764200e-01
#> [101] -1.864089e-01 -1.864089e-01 -1.864089e-01 -1.864089e-01 -1.776855e-01
#> [106] -2.123608e-01 -2.123608e-01 -2.030311e-01 -2.090298e-01 -2.090298e-01
#> [111] -1.944388e-01 -1.631081e-01 -1.471457e-01 -1.471457e-01 -1.549496e-01
#> [116] -1.549496e-01 -1.549496e-01 -1.549496e-01 -1.732543e-01 -1.732543e-01
#> [121] -1.978533e-01 -1.978533e-01 -1.978533e-01 -1.978533e-01 -1.978533e-01
#> [126] -2.056572e-01 -2.056572e-01 -1.939814e-01 -1.939814e-01 -1.939814e-01
#> [131] -1.565066e-01 -1.565066e-01 -1.509395e-01 -1.509395e-01 -1.072350e-01
#> [136] -9.996412e-02 -1.245631e-01 -1.099721e-01 -1.253738e-01 -1.313725e-01
#> [141] -1.313725e-01 -1.313725e-01 -2.075477e-01 -2.075477e-01 -2.200889e-01
#> [146] -2.354905e-01 -2.354905e-01 -2.354905e-01 -2.354905e-01 -2.354905e-01
#> [151] -2.307144e-01 -2.307144e-01 -2.307144e-01 -2.251473e-01 -2.251473e-01
#> [156] -2.195802e-01 -2.140130e-01 -2.140130e-01 -2.067421e-01 -1.973910e-01
#> [161] -4.146238e-01 -3.854228e-01 -3.674452e-01 -3.674452e-01 -3.674452e-01
#> [166] -3.828469e-01 -3.828469e-01 -3.796943e-01 -3.796943e-01 -3.651033e-01
#> [171] -3.651033e-01 -3.882546e-01 -3.840447e-01 -3.840447e-01 -5.485437e-01
#> [176] -5.485437e-01 -5.845810e-01 -5.845810e-01 -6.425615e-01 -6.279705e-01
#> [181] -6.156845e-01 -5.843538e-01 -5.843538e-01 -5.615425e-01 -5.604234e-01
#> [186] -6.132387e-01 -6.132387e-01 -5.818528e-01 -5.818528e-01 -5.663675e-01
#> [191] -5.401419e-01 -5.401419e-01 -5.145774e-01 -5.122344e-01 -5.122344e-01
#> [196] -5.004281e-01 -4.706723e-01 -5.115774e-01 -4.876688e-01 -4.789454e-01
#> [201] -4.772141e-01 -4.655382e-01 -4.953770e-01 -4.837012e-01 -4.837012e-01
#> [206] -4.837012e-01 -4.830177e-01 -4.733948e-01 -5.032336e-01 -4.719029e-01
#> [211] -4.719029e-01 -3.942319e-01 -3.942319e-01 -4.216155e-01 -3.960510e-01
#> [216] -3.800886e-01 -3.758787e-01 -3.706429e-01 -3.706429e-01 -3.209317e-01
#> [221] -3.049693e-01 -2.810194e-01 -2.810194e-01 -2.631128e-01 -2.631128e-01
#> [226] -2.575457e-01 -2.138412e-01 -1.935419e-01 -1.622111e-01 -1.208166e-01
#> [231] -1.155941e-01 -1.001088e-01 -6.189748e-02 -4.117013e-02 -3.639402e-02
#> [236] -2.928015e-02 -2.722049e-02 -2.244438e-02 -4.090693e-02  7.251616e-03
#> [241] -5.289577e-03  9.301381e-03  9.301381e-03  2.691138e-02  2.691138e-02
#> [246]  2.691138e-02 -6.443275e-02 -3.114296e-02 -6.098175e-02 -6.098175e-02
#> [251] -6.098175e-02  6.456976e-02  6.738031e-02  7.261607e-02  7.261607e-02
#> [256]  9.052272e-02  1.021985e-01  1.503571e-01  1.649480e-01  1.766239e-01
#> [261]  1.938910e-01  1.967016e-01  2.060527e-01  2.739827e-01  3.221413e-01
#> [266]  3.314924e-01  2.226461e-01  2.663506e-01  2.193015e-01  2.432514e-01
#> [271]  2.526664e-01  2.547260e-01  2.600998e-01  2.629241e-01  2.775150e-01
#> [276]  3.278100e-01  3.004264e-01  3.068984e-01  3.187047e-01  3.449303e-01
#> [281]  2.869498e-01  2.987562e-01  2.987562e-01  3.163662e-01  3.232744e-01
#> [286]  3.259692e-01  3.285419e-01  3.185530e-01  3.185530e-01  3.222468e-01
#> [291]  2.477334e-01  2.680328e-01  2.732686e-01  2.865609e-01  3.072882e-01
#> [296]  3.454995e-01  3.465910e-01  3.480214e-01  3.635066e-01  4.286349e-01
#> [301]  3.877299e-01  4.116385e-01  3.990973e-01  4.022498e-01  3.260747e-01
#> [306]  3.299411e-01  3.299411e-01  3.445321e-01  3.171485e-01  3.171485e-01
#> [311]  2.939973e-01  2.978637e-01  2.795590e-01  1.822887e-01  1.875245e-01
#> [316]  1.875245e-01  1.880236e-01  1.932594e-01  1.932594e-01  1.170842e-01
#> [321] -1.373697e-02 -1.973569e-02 -1.973569e-02 -1.246477e-02 -9.312201e-03
#> [326]  2.494129e-03 -9.477621e-02 -8.745865e-02 -1.068306e-01 -1.047392e-01
#> [331] -9.917204e-02 -9.917204e-02 -9.270006e-02 -7.810911e-02 -8.809801e-02
#> [336] -9.808692e-02 -8.580092e-02 -8.056515e-02 -2.526061e-01 -2.500334e-01
#> [341] -2.449772e-01 -2.435468e-01 -1.839354e-01 -1.839354e-01 -1.812405e-01
#> [346] -1.689545e-01 -1.647446e-01 -1.945834e-01 -1.772933e-01 -2.092123e-01
#> [351] -2.092123e-01 -1.829867e-01 -2.045666e-01 -2.127774e-01 -2.253186e-01
#> [356] -2.156957e-01 -2.156957e-01 -1.949684e-01 -1.894012e-01 -1.654927e-01
#> [361] -1.636586e-01 -1.618244e-01 -1.442144e-01 -1.442144e-01 -1.421153e-01
#> [366] -1.421153e-01 -1.323020e-01 -1.313302e-01 -1.214023e-01 -9.749378e-02
#> [371] -9.048435e-02 -7.452196e-02 -3.531358e-02 -3.002469e-02 -2.581476e-02
#> [376]  1.239656e-02  1.971412e-02  7.332032e-02  7.332032e-02  4.135634e-02
#> [381]  9.462582e-02  3.666757e-02  3.800287e-02  3.800287e-02  7.955500e-02
#> [386]  1.572260e-01  1.701662e-01  2.016167e-01  2.325988e-01  2.427419e-01
#> [391]  2.464357e-01  2.610266e-01  2.139776e-01  2.285685e-01  2.402444e-01
#> [396]  2.884029e-01  3.004322e-01  3.010814e-01  3.027670e-01  3.092390e-01
#> [401]  3.243608e-01  3.246104e-01  3.766770e-01  3.810328e-01  3.810328e-01
#> [406]  3.986428e-01  4.104491e-01  4.360136e-01  4.413873e-01  4.464435e-01
#> [411]  4.310418e-01  4.439821e-01  4.474927e-01  4.573060e-01  4.573060e-01
#> [416]  4.812146e-01  3.791301e-01  2.353108e-01  1.773526e-01  1.826414e-01
#> [421]  1.899590e-01  1.924768e-01  2.057691e-01  2.296776e-01  2.413535e-01
#> [426]  2.795648e-01  1.015928e-01  1.133992e-01  6.471609e-02  7.343953e-02
#> [431]  6.089834e-02  9.418813e-02  1.255189e-01  1.325283e-01  1.350461e-01
#> [436]  1.379313e-01  1.618812e-01  1.746690e-01  1.892600e-01  1.915748e-01
#> [441]  1.984830e-01  2.673697e-01  3.087642e-01  3.404989e-01  3.404989e-01
#> [446]  3.667245e-01  3.675342e-01  3.880084e-01  3.978217e-01  4.017801e-01
#> [451]  4.105036e-01  4.268177e-01  3.859127e-01  4.267199e-01  4.267199e-01
#> [456]  4.542861e-01  4.014708e-01  4.042951e-01  4.579013e-01  4.677146e-01
#> [461]  4.831999e-01  4.831999e-01  5.269043e-01  4.295425e-01  2.042164e-01
#> [466]  2.129399e-01  2.177160e-01  2.300020e-01  2.429423e-01  2.522720e-01
#> [471]  2.820278e-01  3.023272e-01  3.262357e-01  3.435028e-01  3.508204e-01
#> [476]  3.535152e-01  3.553493e-01  3.570349e-01  3.470460e-01  3.542613e-01
#> [481]  3.326814e-01  3.443572e-01  3.475098e-01  3.788958e-01  4.171071e-01
#> [486]  4.585016e-01  4.590304e-01  4.603657e-01  4.638191e-01  4.845464e-01
#> [491]  4.991374e-01  5.195367e-01  5.210664e-01  5.239517e-01  5.292406e-01
#> [496]  5.317583e-01  5.370472e-01  5.387328e-01  5.233312e-01  5.332591e-01
#> [501]  5.361079e-01  5.368655e-01  5.368655e-01  5.410754e-01  3.192283e-01
#> [506]  3.204551e-01  3.204551e-01  3.408545e-01  3.525303e-01  3.659335e-01
#> [511]  3.659335e-01  3.887448e-01  3.896374e-01  3.899400e-01  4.059024e-01
#> [516]  4.061513e-01  4.090277e-01  4.184426e-01  4.236784e-01  4.534342e-01
#> [521]  3.561638e-01  3.695671e-01  3.695671e-01  3.764753e-01  4.062311e-01
#> [526]  4.182603e-01  4.598124e-01  4.613006e-01  4.638184e-01  4.811085e-01
#> [531]  4.987185e-01  5.007781e-01  5.009259e-01  5.023000e-01  5.122279e-01
#> [536]  5.256311e-01  5.311982e-01  5.444905e-01  5.444905e-01  5.517058e-01
#> [541]  5.552164e-01  5.639398e-01  6.120983e-01  5.784756e-01  6.012868e-01
#> [546]  6.063430e-01  6.081771e-01  6.134660e-01  6.362772e-01  6.958887e-01
#> [551]  6.965362e-01  7.453083e-01  7.495182e-01  7.086132e-01  4.436989e-01
#> [556]  4.553747e-01  4.553747e-01  4.622829e-01  4.666387e-01  4.672862e-01
#> [561]  4.769091e-01  4.842267e-01  4.930630e-01  4.999712e-01  5.025646e-01
#> [566]  4.364708e-01  4.592821e-01  4.631502e-01  4.727731e-01  4.759499e-01
#> [571]  4.803057e-01  4.897207e-01  4.928946e-01  4.998028e-01  5.226140e-01
#> [576]  5.238409e-01  5.249539e-01  5.582437e-01  5.699196e-01  5.903189e-01
#> [581]  6.049099e-01  6.049099e-01  6.101988e-01  6.198216e-01  6.460472e-01
#> [586]  6.487420e-01  5.826482e-01  5.870040e-01  5.997919e-01  6.000945e-01
#> [591]  6.082706e-01  6.200769e-01  6.253658e-01  6.347808e-01  6.379576e-01
#> [596]  6.419160e-01  5.839356e-01  5.892245e-01  5.707619e-01  5.083445e-01
#> [601]  4.809609e-01  4.815498e-01  5.132845e-01  5.158779e-01  5.177121e-01
#> [606]  5.192002e-01  5.251651e-01  5.257376e-01  3.819183e-01  4.023176e-01
#> [611]  4.117326e-01  4.154264e-01  4.154264e-01  4.416520e-01  4.426516e-01
#> [616]  4.495599e-01  4.734684e-01  5.032242e-01  5.061094e-01  5.079436e-01
#> [621]  5.091675e-01  5.120439e-01  5.120439e-01  5.275292e-01  5.503404e-01
#> [626]  5.731517e-01  5.731517e-01  5.738788e-01  3.402401e-01  3.420387e-01
#> [631]  3.455985e-01  3.557416e-01  3.563142e-01  3.802227e-01  3.922520e-01
#> [636]  3.982168e-01  4.046895e-01  4.051178e-01  4.326840e-01  4.920818e-01
#> [641]  4.992971e-01  5.127003e-01  5.355115e-01  5.355115e-01  5.376130e-01
#> [646]  5.498990e-01  5.572165e-01  5.775159e-01  5.590533e-01  5.605415e-01
#> [651]  5.606821e-01  5.779492e-01  5.781981e-01  6.470847e-01  6.562331e-01
#> [656]  6.563560e-01  6.651924e-01  6.677101e-01  7.009999e-01  7.022737e-01
#> [661]  7.045850e-01  6.747462e-01  7.283524e-01  7.355677e-01  7.355677e-01
#> [666]  7.408566e-01  7.254549e-01  7.038750e-01  6.458945e-01  6.613798e-01
#> [671]  6.613798e-01  6.816792e-01  6.976416e-01  7.183689e-01  7.183689e-01
#> [676]  7.218795e-01  3.703972e-01  4.017279e-01  4.046043e-01  1.396900e-01
#> [681]  1.438999e-01  1.491357e-01  1.564066e-01  1.634160e-01  1.687898e-01
#> [686]  1.708912e-01  1.493113e-01  1.493113e-01  5.796714e-02 -1.340323e-02
#> [691]  0.000000e+00
#> 
#> $u
#>   [1] 0.001447178 0.002894356 0.004341534 0.005788712 0.007235890 0.008683068
#>   [7] 0.010130246 0.011577424 0.013024602 0.014471780 0.015918958 0.017366136
#>  [13] 0.018813314 0.020260492 0.021707670 0.023154848 0.024602026 0.026049204
#>  [19] 0.027496382 0.028943560 0.030390738 0.031837916 0.033285094 0.034732272
#>  [25] 0.036179450 0.037626628 0.039073806 0.040520984 0.041968162 0.043415340
#>  [31] 0.044862518 0.046309696 0.047756874 0.049204052 0.050651230 0.052098408
#>  [37] 0.053545586 0.054992764 0.056439942 0.057887120 0.059334298 0.060781476
#>  [43] 0.062228654 0.063675832 0.065123010 0.066570188 0.068017366 0.069464544
#>  [49] 0.070911722 0.072358900 0.073806078 0.075253256 0.076700434 0.078147612
#>  [55] 0.079594790 0.081041968 0.082489146 0.083936324 0.085383502 0.086830680
#>  [61] 0.088277858 0.089725036 0.091172214 0.092619392 0.094066570 0.095513748
#>  [67] 0.096960926 0.098408104 0.099855282 0.101302460 0.102749638 0.104196816
#>  [73] 0.105643994 0.107091172 0.108538350 0.109985528 0.111432706 0.112879884
#>  [79] 0.114327062 0.115774240 0.117221418 0.118668596 0.120115774 0.121562952
#>  [85] 0.123010130 0.124457308 0.125904486 0.127351664 0.128798842 0.130246020
#>  [91] 0.131693198 0.133140376 0.134587554 0.136034732 0.137481910 0.138929088
#>  [97] 0.140376266 0.141823444 0.143270622 0.144717800 0.146164978 0.147612156
#> [103] 0.149059334 0.150506512 0.151953690 0.153400868 0.154848046 0.156295224
#> [109] 0.157742402 0.159189580 0.160636758 0.162083936 0.163531114 0.164978292
#> [115] 0.166425470 0.167872648 0.169319826 0.170767004 0.172214182 0.173661360
#> [121] 0.175108538 0.176555716 0.178002894 0.179450072 0.180897250 0.182344428
#> [127] 0.183791606 0.185238784 0.186685962 0.188133140 0.189580318 0.191027496
#> [133] 0.192474674 0.193921852 0.195369030 0.196816208 0.198263386 0.199710564
#> [139] 0.201157742 0.202604920 0.204052098 0.205499276 0.206946454 0.208393632
#> [145] 0.209840810 0.211287988 0.212735166 0.214182344 0.215629522 0.217076700
#> [151] 0.218523878 0.219971056 0.221418234 0.222865412 0.224312590 0.225759768
#> [157] 0.227206946 0.228654124 0.230101302 0.231548480 0.232995658 0.234442836
#> [163] 0.235890014 0.237337192 0.238784370 0.240231548 0.241678726 0.243125904
#> [169] 0.244573082 0.246020260 0.247467438 0.248914616 0.250361795 0.251808973
#> [175] 0.253256151 0.254703329 0.256150507 0.257597685 0.259044863 0.260492041
#> [181] 0.261939219 0.263386397 0.264833575 0.266280753 0.267727931 0.269175109
#> [187] 0.270622287 0.272069465 0.273516643 0.274963821 0.276410999 0.277858177
#> [193] 0.279305355 0.280752533 0.282199711 0.283646889 0.285094067 0.286541245
#> [199] 0.287988423 0.289435601 0.290882779 0.292329957 0.293777135 0.295224313
#> [205] 0.296671491 0.298118669 0.299565847 0.301013025 0.302460203 0.303907381
#> [211] 0.305354559 0.306801737 0.308248915 0.309696093 0.311143271 0.312590449
#> [217] 0.314037627 0.315484805 0.316931983 0.318379161 0.319826339 0.321273517
#> [223] 0.322720695 0.324167873 0.325615051 0.327062229 0.328509407 0.329956585
#> [229] 0.331403763 0.332850941 0.334298119 0.335745297 0.337192475 0.338639653
#> [235] 0.340086831 0.341534009 0.342981187 0.344428365 0.345875543 0.347322721
#> [241] 0.348769899 0.350217077 0.351664255 0.353111433 0.354558611 0.356005789
#> [247] 0.357452967 0.358900145 0.360347323 0.361794501 0.363241679 0.364688857
#> [253] 0.366136035 0.367583213 0.369030391 0.370477569 0.371924747 0.373371925
#> [259] 0.374819103 0.376266281 0.377713459 0.379160637 0.380607815 0.382054993
#> [265] 0.383502171 0.384949349 0.386396527 0.387843705 0.389290883 0.390738061
#> [271] 0.392185239 0.393632417 0.395079595 0.396526773 0.397973951 0.399421129
#> [277] 0.400868307 0.402315485 0.403762663 0.405209841 0.406657019 0.408104197
#> [283] 0.409551375 0.410998553 0.412445731 0.413892909 0.415340087 0.416787265
#> [289] 0.418234443 0.419681621 0.421128799 0.422575977 0.424023155 0.425470333
#> [295] 0.426917511 0.428364689 0.429811867 0.431259045 0.432706223 0.434153401
#> [301] 0.435600579 0.437047757 0.438494935 0.439942113 0.441389291 0.442836469
#> [307] 0.444283647 0.445730825 0.447178003 0.448625181 0.450072359 0.451519537
#> [313] 0.452966715 0.454413893 0.455861071 0.457308249 0.458755427 0.460202605
#> [319] 0.461649783 0.463096961 0.464544139 0.465991317 0.467438495 0.468885673
#> [325] 0.470332851 0.471780029 0.473227207 0.474674385 0.476121563 0.477568741
#> [331] 0.479015919 0.480463097 0.481910275 0.483357453 0.484804631 0.486251809
#> [337] 0.487698987 0.489146165 0.490593343 0.492040521 0.493487699 0.494934877
#> [343] 0.496382055 0.497829233 0.499276411 0.500723589 0.502170767 0.503617945
#> [349] 0.505065123 0.506512301 0.507959479 0.509406657 0.510853835 0.512301013
#> [355] 0.513748191 0.515195369 0.516642547 0.518089725 0.519536903 0.520984081
#> [361] 0.522431259 0.523878437 0.525325615 0.526772793 0.528219971 0.529667149
#> [367] 0.531114327 0.532561505 0.534008683 0.535455861 0.536903039 0.538350217
#> [373] 0.539797395 0.541244573 0.542691751 0.544138929 0.545586107 0.547033285
#> [379] 0.548480463 0.549927641 0.551374819 0.552821997 0.554269175 0.555716353
#> [385] 0.557163531 0.558610709 0.560057887 0.561505065 0.562952243 0.564399421
#> [391] 0.565846599 0.567293777 0.568740955 0.570188133 0.571635311 0.573082489
#> [397] 0.574529667 0.575976845 0.577424023 0.578871201 0.580318379 0.581765557
#> [403] 0.583212735 0.584659913 0.586107091 0.587554269 0.589001447 0.590448625
#> [409] 0.591895803 0.593342981 0.594790159 0.596237337 0.597684515 0.599131693
#> [415] 0.600578871 0.602026049 0.603473227 0.604920405 0.606367583 0.607814761
#> [421] 0.609261939 0.610709117 0.612156295 0.613603473 0.615050651 0.616497829
#> [427] 0.617945007 0.619392185 0.620839363 0.622286541 0.623733719 0.625180897
#> [433] 0.626628075 0.628075253 0.629522431 0.630969609 0.632416787 0.633863965
#> [439] 0.635311143 0.636758321 0.638205499 0.639652677 0.641099855 0.642547033
#> [445] 0.643994211 0.645441389 0.646888567 0.648335745 0.649782923 0.651230101
#> [451] 0.652677279 0.654124457 0.655571635 0.657018813 0.658465991 0.659913169
#> [457] 0.661360347 0.662807525 0.664254703 0.665701881 0.667149059 0.668596237
#> [463] 0.670043415 0.671490593 0.672937771 0.674384949 0.675832127 0.677279305
#> [469] 0.678726483 0.680173661 0.681620839 0.683068017 0.684515195 0.685962373
#> [475] 0.687409551 0.688856729 0.690303907 0.691751085 0.693198263 0.694645441
#> [481] 0.696092619 0.697539797 0.698986975 0.700434153 0.701881331 0.703328509
#> [487] 0.704775687 0.706222865 0.707670043 0.709117221 0.710564399 0.712011577
#> [493] 0.713458755 0.714905933 0.716353111 0.717800289 0.719247467 0.720694645
#> [499] 0.722141823 0.723589001 0.725036179 0.726483357 0.727930535 0.729377713
#> [505] 0.730824891 0.732272069 0.733719247 0.735166425 0.736613603 0.738060781
#> [511] 0.739507959 0.740955137 0.742402315 0.743849493 0.745296671 0.746743849
#> [517] 0.748191027 0.749638205 0.751085384 0.752532562 0.753979740 0.755426918
#> [523] 0.756874096 0.758321274 0.759768452 0.761215630 0.762662808 0.764109986
#> [529] 0.765557164 0.767004342 0.768451520 0.769898698 0.771345876 0.772793054
#> [535] 0.774240232 0.775687410 0.777134588 0.778581766 0.780028944 0.781476122
#> [541] 0.782923300 0.784370478 0.785817656 0.787264834 0.788712012 0.790159190
#> [547] 0.791606368 0.793053546 0.794500724 0.795947902 0.797395080 0.798842258
#> [553] 0.800289436 0.801736614 0.803183792 0.804630970 0.806078148 0.807525326
#> [559] 0.808972504 0.810419682 0.811866860 0.813314038 0.814761216 0.816208394
#> [565] 0.817655572 0.819102750 0.820549928 0.821997106 0.823444284 0.824891462
#> [571] 0.826338640 0.827785818 0.829232996 0.830680174 0.832127352 0.833574530
#> [577] 0.835021708 0.836468886 0.837916064 0.839363242 0.840810420 0.842257598
#> [583] 0.843704776 0.845151954 0.846599132 0.848046310 0.849493488 0.850940666
#> [589] 0.852387844 0.853835022 0.855282200 0.856729378 0.858176556 0.859623734
#> [595] 0.861070912 0.862518090 0.863965268 0.865412446 0.866859624 0.868306802
#> [601] 0.869753980 0.871201158 0.872648336 0.874095514 0.875542692 0.876989870
#> [607] 0.878437048 0.879884226 0.881331404 0.882778582 0.884225760 0.885672938
#> [613] 0.887120116 0.888567294 0.890014472 0.891461650 0.892908828 0.894356006
#> [619] 0.895803184 0.897250362 0.898697540 0.900144718 0.901591896 0.903039074
#> [625] 0.904486252 0.905933430 0.907380608 0.908827786 0.910274964 0.911722142
#> [631] 0.913169320 0.914616498 0.916063676 0.917510854 0.918958032 0.920405210
#> [637] 0.921852388 0.923299566 0.924746744 0.926193922 0.927641100 0.929088278
#> [643] 0.930535456 0.931982634 0.933429812 0.934876990 0.936324168 0.937771346
#> [649] 0.939218524 0.940665702 0.942112880 0.943560058 0.945007236 0.946454414
#> [655] 0.947901592 0.949348770 0.950795948 0.952243126 0.953690304 0.955137482
#> [661] 0.956584660 0.958031838 0.959479016 0.960926194 0.962373372 0.963820550
#> [667] 0.965267728 0.966714906 0.968162084 0.969609262 0.971056440 0.972503618
#> [673] 0.973950796 0.975397974 0.976845152 0.978292330 0.979739508 0.981186686
#> [679] 0.982633864 0.984081042 0.985528220 0.986975398 0.988422576 0.989869754
#> [685] 0.991316932 0.992764110 0.994211288 0.995658466 0.997105644 0.998552822
#> [691] 1.000000000
#> 
#> $covariate
#> [1] "reciprocity_count"
#> 
# }